chore: audit remediation — fixes, docs, spec, SPM, tooling
Static audit (agent-os/specs/2026-04-30-1845-project-audit) produced a
remediation list of code defects, doc drift, spec staleness, and dependency
hygiene. This commit lands the agreed scope.

Code fixes (with tests):
- :argus-urlconnection — fix body discard when disconnect() runs before
  close(); TeeInputStream now exposes capturedBytes/totalBytesRead so
  emitIfNeeded() drains the live tee. Adds regression test.
- :argus-okhttp — replace full request-body materialization with a
  streaming TeeingRequestBody (ForwardingSink + Buffer.copyTo). Memory
  is bounded by maxBodyBytes, not body size. Adds StreamingRequestBodyTest.
- argus-core BodyCapture.drainWithCap — drop ArrayList<Byte> per-byte
  boxing in favor of primitive ByteArray + grow strategy.
- effectiveMaxBytesFor — clamp Long.MAX_VALUE to Int.MAX_VALUE so
  fullBodyHosts truncation surfaces honestly via truncatedTotalBytes;
  delete the duplicate private helper in ArgusClientPlugin and route all
  three engines through ArgusClientConfig.toCaptureConfig().
- expect/actual KClass<*>.bestEffortFqn() — JVM/Android return
  qualifiedName, iOS falls back to simpleName. Replaces 4 capture sites
  that used simpleName only, distinguishing same-name exception classes
  on JVM.
- ArgusHandle.startupError: StateFlow<Throwable?> on Android+iOS;
  Argus.start() wraps the launched coroutine in try/catch so embedded
  server bind failures surface instead of silently leaving url=null.
- ArgusSmokeTest for both platforms (Robolectric on Android, runBlocking
  on iOS); iOS test binary now links -lsqlite3 explicitly.

UI:
- BodyViewer drops the dead "Load full body" button (forbidden by spec
  §5.7; no caller wired it).
- New fontSize.xxs token (10px/14px) added to design handoff CSS;
  three text-[10px] arbitrary-Tailwind sites switched to text-xxs.

iOS distribution (XCFramework + SPM):
- argus-ios/build.gradle.kts wires XCFramework("argus-ios"); test binaries
  pin -lsqlite3.
- New Package.swift at repo root with binaryTarget pointing at GitHub
  Releases. Per-release URL/checksum updated manually for now.
- publishToMavenCentral.yml extended: build XCFramework, zip, compute
  SHA-256, upload as a release asset, and append the SPM
  URL+checksum to the release body.

README + spec hygiene:
- README §1 intro adds the Argus Panoptes name origin.
- §9 Configuration reference split into 9.1 server + 9.2 client tables;
  persist trio + client capture knobs documented.
- §4/§5 sample blocks now truly verbatim with :sample (all 7 DebugTools
  methods on Android and iOS impls).
- §11 module table + Mermaid diagram extended with argus-ios,
  argus-okhttp, argus-urlconnection. SPM install path documented in §5.
- KDoc on Argus/ArgusHandle (Android+iOS), Ktor Argus plugin val, and
  ArgusClientConfig properties.
- argus-audit-spec.md: design refs repointed at
  design_handoff_argus_inspector/, server tests acknowledged as jvmTest,
  Waterfall description matches impl, keyboard x/Shift+X amended,
  ArgusServer plain-class decision documented, :sample-android renamed
  to :sample throughout, §7 phase-deferral list replaced with
  "Implementation status" reflecting MVP shipped state.
- agent-os/product/mission.md: stale :lantern-android / mDNS claim
  rewritten to match the spec §3.8 reality.
- Strip 37 stale "Phase N" mentions across argus-core, argus-server-core,
  argus-webui, sample DebugTools KDoc, and the spec.

Tooling:
- argus-webui: vite ^6, vitest ^3, tsx ^4, postcss ^8.5.13. Clears the
  vite/esbuild dev-server CVEs flagged by the audit; npm install reports
  zero vulnerabilities; build + tests pass on the new majors.

Deferred (intentional, documented in plan):
- R9 ring-buffer snapshot allocation: needs perf measurement (LoadTest)
  before changing the public StateFlow<List<ArgusEvent>> API.
- R12 WS subscribe ack: hypothetical impact, web UI does client-side
  filtering.
- R15 sample minSdk 34 pin: already inline-documented in
  sample/build.gradle.kts.

+++ argus-core/src/commonMain/kotlin/com/lynxal/argus/capture/CaptureConfig.kt
@@ -13,5 +13,8 @@ public data class CaptureConfig(
 public fun effectiveMaxBytesFor(host: String, cfg: CaptureConfig): Long {
     if (cfg.fullBodyHosts.isEmpty()) return cfg.maxBodyBytes
     val match = cfg.fullBodyHosts.any { it.equals(host, ignoreCase = true) }
-    return if (match) Long.MAX_VALUE else cfg.maxBodyBytes
+    // Clamp to Int.MAX_VALUE: captured bodies live in a single ByteArray, which
+    // is Int-sized. Returning Long.MAX_VALUE would silently truncate at ~2 GB
+    // without `truncatedTotalBytes` reflecting the cap.
+    return if (match) Int.MAX_VALUE.toLong() else cfg.maxBodyBytes
 }
